+++ packages/ragbits-agents/src/ragbits/agents/_main.py
@@ -10,6 +10,7 @@ from typing import ClassVar, Generic, cast, overload
 from ragbits import agents
 from ragbits.agents.exceptions import (
     AgentInvalidPromptInputError,
+    AgentMaxTurnsExceededError,
     AgentToolDuplicateError,
     AgentToolExecutionError,
     AgentToolNotAvailableError,
@@ -54,6 +55,9 @@ class AgentOptions(Options, Generic[LLMClientOptionsT]):
 
     llm_options: LLMClientOptionsT | None | NotGiven = NOT_GIVEN
     """The options for the LLM."""
+    max_turns: int | None | NotGiven = NOT_GIVEN
+    """The maximum number of turns the agent can take, if NOT_GIVEN,
+    it defaults to 10, if None, agent will run forever"""
 
 
 class AgentResultStreaming(AsyncIterator[str | ToolCall | ToolCallResult]):
@@ -187,9 +191,9 @@ class Agent(
             AgentToolNotSupportedError: If the selected tool type is not supported.
             AgentToolNotAvailableError: If the selected tool is not available.
             AgentInvalidPromptInputError: If the prompt/input combination is invalid.
+            AgentMaxTurnsExceededError: If the maximum number of turns is exceeded.
         """
         input = cast(PromptInputT, input)
-
         merged_options = (self.default_options | options) if options else self.default_options
         llm_options = merged_options.llm_options or None
 
@@ -197,8 +201,11 @@ class Agent(
         tools_mapping = await self._get_all_tools()
         tool_calls = []
 
+        turn_count = 0
+        max_turns = merged_options.max_turns
+        max_turns = 10 if max_turns is NOT_GIVEN else max_turns
         with trace(input=input, options=merged_options) as outputs:
-            while True:
+            while not max_turns or turn_count < max_turns:
                 response = cast(
                     LLMResponseWithMetadata[PromptOutputT],
                     await self.llm.generate_with_metadata(
@@ -215,6 +222,10 @@ class Agent(
                     tool_calls.append(result)
 
                     prompt_with_history = prompt_with_history.add_tool_use_message(**result.__dict__)
+
+                turn_count += 1
+            else:
+                raise AgentMaxTurnsExceededError(cast(int, max_turns))
 
             outputs.result = {
                 "content": response.content,
@@ -267,6 +278,7 @@ class Agent(
             AgentToolNotSupportedError: If the selected tool type is not supported.
             AgentToolNotAvailableError: If the selected tool is not available.
             AgentInvalidPromptInputError: If the prompt/input combination is invalid.
+            AgentMaxTurnsExceededError: If the maximum number of turns is exceeded.
         """
         generator = self._stream_internal(input, options)
         return AgentResultStreaming(generator)
@@ -280,9 +292,11 @@ class Agent(
 
         prompt_with_history = self._get_prompt_with_history(input)
         tools_mapping = await self._get_all_tools()
-
+        turn_count = 0
+        max_turns = merged_options.max_turns
+        max_turns = 10 if max_turns is NOT_GIVEN else max_turns
         with trace(input=input, options=merged_options) as outputs:
-            while True:
+            while not max_turns or turn_count < max_turns:
                 returned_tool_call = False
                 async for chunk in self.llm.generate_streaming(
                     prompt=prompt_with_history,
@@ -296,9 +310,13 @@ class Agent(
                         yield result
                         prompt_with_history = prompt_with_history.add_tool_use_message(**result.__dict__)
                         returned_tool_call = True
+                turn_count += 1
 
                 if not returned_tool_call:
                     break
+            else:
+                raise AgentMaxTurnsExceededError(cast(int, max_turns))
+
             yield prompt_with_history
             if self.keep_history:
                 self.history = prompt_with_history.chat