+++ typescript/@ragbits/api-client-react/__tests__/utils.tsx
@@ -6,7 +6,7 @@ import {
     RenderHookOptions,
     RenderResult,
 } from '@testing-library/react'
-import { RagbitsProvider } from '../src'
+import { RagbitsContextProvider } from '../src'
 import type { ConfigResponse } from '@ragbits/api-client'
 
 interface CustomRenderOptions extends Omit<RenderOptions, 'wrapper'> {
@@ -20,7 +20,11 @@ interface CustomRenderHookOptions<TProps>
 
 function createWrapper(baseUrl?: string) {
     return function Wrapper({ children }: { children: React.ReactNode }) {
-        return <RagbitsProvider baseUrl={baseUrl}>{children}</RagbitsProvider>
+        return (
+            <RagbitsContextProvider baseUrl={baseUrl}>
+                {children}
+            </RagbitsContextProvider>
+        )
     }
 }
 
@@ -111,5 +115,5 @@ export const defaultConfigResponse: ConfigResponse = {
             },
         },
     },
-    debug_mode: true,
+    debug_mode: false,
 }